+++ function/cross_entropy_test.go
@@ -48,7 +48,7 @@ func ExampleOneHot() {
 }
 
 func ExampleLogp() {
-	x := tensor.New([]int{3, 5}, []float64{
+	x := tensor.New([]int{3, 5}, []float32{
 		1, 2, 3, 4, 5,
 		6, 7, 8, 9, 10,
 		11, 12, 13, 14, 15,
